Split time span element util functions to their own file
