+++ tests/test_external_services/test_hauki/test_override_reservable_with_closed_time_spans.py
@@ -1,5 +1,5 @@
-from opening_hours.utils.reservable_time_span_client import override_reservable_with_closed_time_spans
 from opening_hours.utils.time_span_element import TimeSpanElement
+from opening_hours.utils.time_span_element_utils import override_reservable_with_closed_time_spans
 from tests.test_external_services.test_hauki.test_reservable_time_spans_client import (
     _get_date,
     _get_normalised_time_spans,
